Api - More Id structure fixes and fixing Driver Create endpoint

+++ Journal-Services/Authentication/Journal.Identity/Models/User/Role.cs
@@ -6,5 +6,5 @@ namespace Journal.Identity.Models.User;
 [UsedImplicitly]
 public class Role : IdentityRole
 {
-    public int SecondaryId { get; set; }
+    public long SecondaryId { get; set; }
 }
